+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
--
+### Added
+
+- Add initial cookies to login function to prevent captcha requests in most cases.
 
 ## [0.5.1] - 2020-01-05